+++ refinedstorage2-network/src/main/java/com/refinedmods/refinedstorage2/api/network/impl/node/container/NetworkNodeContainerPriorities.java
@@ -1,0 +1,8 @@
+package com.refinedmods.refinedstorage2.api.network.impl.node.container;
+
+public final class NetworkNodeContainerPriorities {
+    public static final int GRID = Integer.MAX_VALUE;
+
+    private NetworkNodeContainerPriorities() {
+    }
+}