+++ refinedstorage2-network/src/test/java/com/refinedmods/refinedstorage2/api/network/impl/FakeConnectionProvider.java
@@ -74,7 +74,7 @@ public class FakeConnectionProvider implements ConnectionProvider {
     }
 
     @Override
-    public List<NetworkNodeContainer> sort(final Set<NetworkNodeContainer> containers) {
+    public List<NetworkNodeContainer> sortDeterministically(final Set<NetworkNodeContainer> containers) {
         return containers
             .stream()
             .sorted(Comparator.comparingInt(allowed::indexOf))