+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 -   Fixed missing Speed Upgrade energy usage config on Forge.
+-   Fixed Grid screen not handling network changes properly.
 
 ### Removed